Change start script file target
