+++ package.json
@@ -20,7 +20,7 @@
     "typescript": "^5.2.2"
   },
   "scripts": {
-    "start": "node ./dist/CrossingGuardBot.js",
+    "start": "node ./dist/index.js",
     "tsc": "tsc",
     "postinstall": "npm run tsc",
     "dev": "ts-node-dev --respawn --pretty --transpile-only App.ts",
