+++ src/index.ts
@@ -1,0 +1,3 @@
+import CrossingGuardBot from "./bot/CrossingGuardBot";
+
+var bot = CrossingGuardBot.getInstance();
