+++ src/util/util.ts
@@ -1,0 +1,11 @@
+function ensureError(value: unknown): Error {
+    if (value instanceof Error) return value
+
+    let stringified = '[Unable to stringify the thrown value]'
+    try {
+        stringified = JSON.stringify(value)
+    } catch { }
+
+    const error = new Error(`This value was thrown as is, not through an Error: ${stringified}`)
+    return error
+}
